1-last_digit.c

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -8,7 +8,7 @@
   */
 int main(void)
 {
-	long  n;
+        float  n;
 	int ld;
 
 	srand(time(0));
@@ -16,15 +16,15 @@ int main(void)
 	ld = n % 10;
 	if (ld > 5)
 	{
-		printf("Last digit of %d is %d and is greater than 5\n", n, ld);
+		printf("Last digit of %f is %d and is greater than 5\n", n, ld);
 	}
 	else if (ld == 0)
 	{
-		printf("Last digit of %d is %d and is 0\n", n, ld);
+		printf("Last digit of %f is %d and is 0\n", n, ld);
 	}
 	else
 	{
-		printf("Last digit of %d is %d and less than 6 and not 0\n", n, ld);
+		printf("Last digit of %f is %d and less than 6 and not 0\n", n, ld);
 	}
 	return (0);
 }
